+++ tests/test_osmnx.py
@@ -7,7 +7,6 @@ import matplotlib as mpl
 mpl.use("Agg")
 
 import bz2
-import contextlib
 import logging as lg
 import os
 import tempfile
@@ -343,15 +342,32 @@ def test_find_nearest():
 
 
 def test_endpoints():
+    default_timeout = ox.settings.timeout
+    default_key = ox.settings.nominatim_key
+    default_nominatim_endpoint = ox.settings.nominatim_endpoint
+    default_overpass_endpoint = ox.settings.overpass_endpoint
+    default_overpass_rate_limit = ox.settings.overpass_rate_limit
+
+    # test good and bad DNS resolution
+    ox.settings.timeout = 1
     ip = ox._downloader._resolve_host_via_doh("overpass-api.de")
     ip = ox._downloader._resolve_host_via_doh("AAAAAAAAAAA")
-
     _doh_url_template_default = ox.settings.doh_url_template
     ox.settings.doh_url_template = "http://aaaaaa.hostdoesntexist.org/nothinguseful"
     ip = ox._downloader._resolve_host_via_doh("overpass-api.de")
     ox.settings.doh_url_template = None
     ip = ox._downloader._resolve_host_via_doh("overpass-api.de")
     ox.settings.doh_url_template = _doh_url_template_default
+
+    # Test changing the Overpass endpoint.
+    # This should fail because we didn't provide a valid endpoint
+    ox.settings.overpass_rate_limit = False
+    ox.settings.overpass_endpoint = "http://NOT_A_VALID_ENDPOINT/api/"
+    with pytest.raises(ConnectionError, match="Max retries exceeded with url"):
+        G = ox.graph_from_place(place1, network_type="all")
+
+    ox.settings.overpass_rate_limit = default_overpass_rate_limit
+    ox.settings.timeout = default_timeout
 
     params = OrderedDict()
     params["format"] = "json"
@@ -377,19 +393,9 @@ def test_endpoints():
     with pytest.raises(ValueError, match="Nominatim request_type must be"):
         response_json = ox._downloader._nominatim_request(params=params, request_type="xyz")
 
-    default_key = ox.settings.nominatim_key
-    default_nominatim_endpoint = ox.settings.nominatim_endpoint
-    default_overpass_endpoint = ox.settings.overpass_endpoint
-
-    # Searching on public nominatim should work even if a key was provided
+    # Searching on public nominatim should work even if a (bad) key was provided
     ox.settings.nominatim_key = "NOT_A_KEY"
     response_json = ox._downloader._nominatim_request(params=params, request_type="search")
-
-    # Test changing the endpoint.
-    # This should fail because we didn't provide a valid endpoint
-    ox.settings.overpass_endpoint = "http://NOT_A_VALID_ENDPOINT/api/"
-    with pytest.raises(ConnectionError, match="Max retries exceeded with url"):
-        G = ox.graph_from_place(place1, network_type="all")
 
     ox.settings.nominatim_key = default_key
     ox.settings.nominatim_endpoint = default_nominatim_endpoint
@@ -511,7 +517,7 @@ def test_graph_from_functions():
 
 def test_features():
     # geometries_from_bbox - bounding box query to return no data
-    with contextlib.suppress(ox._errors.EmptyResponseError):
+    with pytest.raises(ox._errors.InsufficientResponseError):
         gdf = ox.geometries_from_bbox(0.009, -0.009, 0.009, -0.009, tags={"building": True})
 
     # geometries_from_bbox - successful